Revert "refactor: extract JobQueue configuration"
This reverts commit 5778936b712befd6e367bee3decc4a3565ca69af.
