+++ Queues/JobQueue.cs
@@ -13,16 +13,21 @@
  * limitations under the License.
 */
 
-using System;
-using System.IO;
+using Fasterflect;
 using Newtonsoft.Json;
-using QuantConnect.Util;
-using QuantConnect.Python;
-using QuantConnect.Packets;
-using QuantConnect.Logging;
-using QuantConnect.Interfaces;
+using QuantConnect.Brokerages;
 using QuantConnect.Configuration;
+using QuantConnect.Data;
+using QuantConnect.Interfaces;
+using QuantConnect.Logging;
+using QuantConnect.Packets;
+using QuantConnect.Python;
+using QuantConnect.Util;
+using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Reflection;
 
 namespace QuantConnect.Queues
 {
@@ -33,6 +38,12 @@ namespace QuantConnect.Queues
     {
         // The type name of the QuantConnect.Brokerages.Paper.PaperBrokerage
         private static readonly TextWriter Console = System.Console.Out;
+
+        private const string PaperBrokerageTypeName = "PaperBrokerage";
+        private const string DefaultHistoryProvider = "SubscriptionDataReaderHistoryProvider";
+        private const string DefaultDataQueueHandler = "LiveDataQueue";
+        private const string DefaultDataChannelProvider = "DataChannelProvider";
+        private static readonly string Channel = Config.Get("data-channel");
         private readonly string AlgorithmTypeName = Config.Get("algorithm-type-name");
         private Language? _language;
 
@@ -86,6 +97,28 @@ namespace QuantConnect.Queues
         }
 
         /// <summary>
+        /// Gets Brokerage Factory for provided IDQH
+        /// </summary>
+        /// <param name="dataQueueHandler"></param>
+        /// <returns>An Instance of Brokerage Factory if possible, otherwise null</returns>
+        public static IBrokerageFactory GetFactoryFromDataQueueHandler(string dataQueueHandler)
+        {
+            IBrokerageFactory brokerageFactory = null;
+            var dataQueueHandlerType = Composer.Instance.GetExportedTypes<IBrokerage>()
+                .FirstOrDefault(x =>
+                    x.FullName != null &&
+                    x.FullName.EndsWith(dataQueueHandler, StringComparison.InvariantCultureIgnoreCase) &&
+                    x.HasAttribute(typeof(BrokerageFactoryAttribute)));
+
+            if (dataQueueHandlerType != null)
+            {
+                var attribute = dataQueueHandlerType.GetCustomAttribute<BrokerageFactoryAttribute>();
+                brokerageFactory = (BrokerageFactory)Activator.CreateInstance(attribute.Type);
+            }
+            return brokerageFactory;
+        }
+
+        /// <summary>
         /// Desktop/Local Get Next Task - Get task from the Algorithm folder of VS Solution.
         /// </summary>
         /// <returns></returns>
@@ -104,17 +137,113 @@ namespace QuantConnect.Queues
                 parameters = JsonConvert.DeserializeObject<Dictionary<string, string>>(parametersConfigString);
             }
 
+            var controls = new Controls()
+            {
+                MinuteLimit = Config.GetInt("symbol-minute-limit", 10000),
+                SecondLimit = Config.GetInt("symbol-second-limit", 10000),
+                TickLimit = Config.GetInt("symbol-tick-limit", 10000),
+                RamAllocation = int.MaxValue,
+                MaximumDataPointsPerChartSeries = Config.GetInt("maximum-data-points-per-chart-series", 1000000),
+                MaximumChartSeries = Config.GetInt("maximum-chart-series", 30),
+                StorageLimit = Config.GetValue("storage-limit", 10737418240L),
+                StorageFileCount = Config.GetInt("storage-file-count", 10000),
+                StoragePermissions = (FileAccess)Config.GetInt("storage-permissions", (int)FileAccess.ReadWrite)
+            };
+
             var algorithmId = Config.Get("algorithm-id", AlgorithmTypeName);
 
             //If this isn't a backtesting mode/request, attempt a live job.
             if (Globals.LiveMode)
             {
-                return JobQueueExtensions.GetLiveNodeConfigurationWithAlgorithmConfiguration(AlgorithmLocation, algorithmId, parameters, Language);
+                var dataHandlers = Config.Get("data-queue-handler", DefaultDataQueueHandler);
+                var liveJob = new LiveNodePacket
+                {
+                    Type = PacketType.LiveNode,
+                    Algorithm = File.ReadAllBytes(AlgorithmLocation),
+                    Brokerage = Config.Get("live-mode-brokerage", PaperBrokerageTypeName),
+                    HistoryProvider = Config.Get("history-provider", DefaultHistoryProvider),
+                    DataQueueHandler = dataHandlers,
+                    DataChannelProvider = Config.Get("data-channel-provider", DefaultDataChannelProvider),
+                    Channel = Channel,
+                    UserToken = Globals.UserToken,
+                    UserId = Globals.UserId,
+                    ProjectId = Globals.ProjectId,
+                    OrganizationId = Globals.OrganizationID,
+                    Version = Globals.Version,
+                    DeployId = algorithmId,
+                    Parameters = parameters,
+                    Language = Language,
+                    Controls = controls,
+                    PythonVirtualEnvironment = Config.Get("python-venv"),
+                    DeploymentTarget = DeploymentTarget.LocalPlatform,
+                };
+
+                Type brokerageName = null;
+                try
+                {
+                    // import the brokerage data for the configured brokerage
+                    var brokerageFactory = Composer.Instance.Single<IBrokerageFactory>(factory => factory.BrokerageType.MatchesTypeName(liveJob.Brokerage));
+                    brokerageName = brokerageFactory.BrokerageType;
+                    liveJob.BrokerageData = brokerageFactory.BrokerageData;
+                }
+                catch (Exception err)
+                {
+                    Log.Error(err, $"Error resolving BrokerageData for live job for brokerage {liveJob.Brokerage}");
+                }
+
+                var brokerageBasedHistoryProvider = liveJob.HistoryProvider.DeserializeList().Select(x =>
+                {
+                    HistoryExtensions.TryGetBrokerageName(x, out var brokerageName);
+                    return brokerageName;
+                }).Where(x => x != null);
+
+                foreach (var dataHandlerName in dataHandlers.DeserializeList().Concat(brokerageBasedHistoryProvider).Distinct())
+                {
+                    var brokerageFactoryForDataHandler = GetFactoryFromDataQueueHandler(dataHandlerName);
+                    if (brokerageFactoryForDataHandler == null)
+                    {
+                        Log.Trace($"JobQueue.NextJob(): Not able to fetch brokerage factory with name: {dataHandlerName}");
+                        continue;
+                    }
+                    if (brokerageFactoryForDataHandler.BrokerageType == brokerageName)
+                    {
+                        //Don't need to add brokearageData again if added by brokerage
+                        continue;
+                    }
+                    foreach (var data in brokerageFactoryForDataHandler.BrokerageData)
+                    {
+                        if (data.Key == "live-holdings" || data.Key == "live-cash-balance")
+                        {
+                            //live holdings & cash balance not required for data handler
+                            continue;
+                        }
+
+                        liveJob.BrokerageData.TryAdd(data.Key, data.Value);
+                    }
+                }
+                return liveJob;
             }
 
             var optimizationId = Config.Get("optimization-id");
             //Default run a backtesting job.
-            var backtestJob = JobQueueExtensions.GetBacktestNodePacketConfiguration(AlgorithmLocation, algorithmId, parameters, Language);
+            var backtestJob = new BacktestNodePacket(0, 0, "", new byte[] { }, Config.Get("backtest-name", "local"))
+            {
+                Type = PacketType.BacktestNode,
+                Algorithm = File.ReadAllBytes(AlgorithmLocation),
+                HistoryProvider = Config.Get("history-provider", DefaultHistoryProvider),
+                Channel = Channel,
+                UserToken = Globals.UserToken,
+                UserId = Globals.UserId,
+                ProjectId = Globals.ProjectId,
+                OrganizationId = Globals.OrganizationID,
+                Version = Globals.Version,
+                BacktestId = algorithmId,
+                Language = Language,
+                Parameters = parameters,
+                Controls = controls,
+                PythonVirtualEnvironment = Config.Get("python-venv"),
+                DeploymentTarget = DeploymentTarget.LocalPlatform,
+            };
 
             var outOfSampleMaxEndDate = Config.Get("out-of-sample-max-end-date");
             if (!string.IsNullOrEmpty(outOfSampleMaxEndDate))
